+++ docs/data/README.md
@@ -9,12 +9,8 @@ Repositorio de datos
 
 ## spat_covid_v2.csv 
 
-<<<<<<< HEAD
-  - Datos del censo - Recopilado por @gcarrascoe
-=======
   - Datos del censo - Recopilado por @gcarrascoe
   
 ## Renipress_2020
 
   - Datos Renipress sobre capacidad hospitalaria- Recopilado por @jjenifer
->>>>>>> f20818b81e46628a0e5c9e7b9ce032b575909b67
